MSI: Handle some int conversion warnings.

+++ src/modules/msi/cab/qtmd.c
@@ -266,7 +266,7 @@ int qtmd_decompress(struct qtmd_stream *qtm, off_t out_bytes) {
   if (qtm->error) return qtm->error;
 
   /* flush out any stored-up bytes before we begin */
-  i = qtm->o_end - qtm->o_ptr;
+  i = (int) (qtm->o_end - qtm->o_ptr);
   if ((off_t) i > out_bytes) i = (int) out_bytes;
   if (i) {
     if (qtm->sys->write(qtm->output, qtm->o_ptr, i) != i) {
@@ -288,7 +288,7 @@ int qtmd_decompress(struct qtmd_stream *qtm, off_t out_bytes) {
 
   /* while we do not have enough decoded bytes in reserve: */
   while ((qtm->o_end - qtm->o_ptr) < out_bytes) {
-    /* read header if necessary. Initialises H, L and C */
+    /* read header if necessary. Initializes H, L and C */
     if (!qtm->header_read) {
       H = 0xFFFF; L = 0; READ_BITS(C, 16);
       qtm->header_read = 1;
@@ -296,7 +296,7 @@ int qtmd_decompress(struct qtmd_stream *qtm, off_t out_bytes) {
 
     /* decode more, up to the number of bytes needed, the frame boundary,
      * or the window boundary, whichever comes first */
-    frame_end = window_posn + (out_bytes - (qtm->o_end - qtm->o_ptr));
+    frame_end = window_posn + (unsigned int) (out_bytes - (qtm->o_end - qtm->o_ptr));
     if ((window_posn + frame_todo) < frame_end) {
       frame_end = window_posn + frame_todo;
     }
@@ -362,7 +362,7 @@ int qtmd_decompress(struct qtmd_stream *qtm, off_t out_bytes) {
 	  while (i--) *rundest++ = window[j++ & (qtm->window_size - 1)];
 
 	  /* flush currently stored data */
-	  i = (&window[qtm->window_size] - qtm->o_ptr);
+	  i = (int) (&window[qtm->window_size] - qtm->o_ptr);
 
 	  /* this should not happen, but if it does then this code
 	   * can't handle the situation (can't flush up to the end of
@@ -444,7 +444,7 @@ int qtmd_decompress(struct qtmd_stream *qtm, off_t out_bytes) {
     /* window wrap? */
     if (window_posn == qtm->window_size) {
       /* flush all currently stored data */
-      i = (qtm->o_end - qtm->o_ptr);
+      i = (int) (qtm->o_end - qtm->o_ptr);
       /* break out if we have more than enough to finish this request */
       if (i >= out_bytes) break;
       if (qtm->sys->write(qtm->output, qtm->o_ptr, i) != i) {
